requirements.txt
requests
opencv-python
python-dotenv
RedisPy
apscheduler
pymongo